+++ src/index.ts
@@ -1,4 +1,7 @@
-// Time related utils
+// Data related utils
+export { default as extractBase64Data } from './utils/data/extractBase64Data.ts'
 
+// Time related utils
 export { default as isTimestampExpired } from './utils/time/isTimestampExpired.ts'
 export { default as convertDateToTimestamp } from './utils/time/convertDateToTimestamp.ts'
+export { default as convertTimestampToDate } from './utils/time/convertTimestampToDate.ts'