+++ CHANGELOG.md
@@ -1,16 +1,22 @@
 # Changelog
 
+### 0.1.3
+
+Release 23 sep 2023
+  * Add [`convertTimestampToDate`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#convertTimestampToDate) to time related utils
+  * Add [`extractBase64Data`](https://github.com/ovniroto/ovni-utils/wiki/Recipebook#extractBase64Data) to data related utils
+
 ### 0.1.2
 
-* Release 22 sep 2023
+Release 22 sep 2023
   * Some build fixes
 
 ### 0.1.1
 
-* Release 22 sep 2023
+Release 22 sep 2023
   * Some changes
 
 ### 0.1.0
 
-* Release 22 sep 2023
+Release 22 sep 2023
   * First version